refactor : 코드 추출
