+++ components/modal/RecordUpdateModal.tsx
@@ -13,48 +13,14 @@ import RNModal from 'react-native-modal';
 import Colors from '@/constants/Colors';
 import Sizes from '@/constants/Sizes';
 import getSize from '@/scripts/getSize';
-import { ModalHeader } from '@/components/modal/UpdateModal';
+import {
+  ModalHeader,
+  ModalInput,
+  ModalProps,
+} from '@/components/modal/UpdateModal';
 
 const { width } = Dimensions.get('window');
 
-// Interface for the common modal props
-interface ModalProps {
-  isVisible?: boolean;
-  onCancel?: () => void;
-  onConfirm?: (newValue: string) => void;
-  value?: string;
-  onChangeText: (text: string) => void;
-  label?: string;
-  placeholder?: string;
-}
-
-// Modal input component
-const ModalInput: React.FC<{
-  label: string;
-  value: string;
-  onChangeText: (text: string) => void;
-  placeholder: string;
-}> = ({
-  label,
-  value,
-  onChangeText,
-  placeholder
-}) => (
-    <View style={styles.modalInputContainer}>
-      <Text style={styles.modalInputTitle}>{label}</Text>
-      <TextInput
-        value={value}
-        onChangeText={onChangeText}
-        style={styles.modalInput}
-        placeholder={placeholder}
-        placeholderTextColor="rgba(255, 255, 255, 0.7)"
-        multiline={true}
-        maxLength={200}
-      />
-    </View>
-  );
-
-// General RecordUpdateModal component that other modals will use
 const RecordUpdateModal: React.FC<ModalProps & { title: string }> = ({
   isVisible,
   onCancel,
@@ -144,53 +110,10 @@ const styles = StyleSheet.create({
     borderTopRightRadius: 20,
     alignItems: 'center',
   },
-  modalTitleContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    width: width,
-    height: getSize(29),
-    paddingHorizontal: getSize(24),
-    marginTop: getSize(13),
-  },
-  modalCancel: {
-    fontSize: getSize(16),
-    fontFamily: 'Pretendard-SemiBold',
-    color: 'white',
-  },
-  modalTitle: {
-    fontSize: getSize(Sizes.pageTitle),
-    fontFamily: 'Pretendard-SemiBold',
-    color: 'white',
-  },
-  modalConfirm: {
-    fontSize: getSize(16),
-    fontFamily: 'Pretendard-SemiBold',
-    color: Colors.main,
-  },
-  modalInputContainer: {
-    paddingLeft: getSize(18),
-    width: width - getSize(Sizes.formMargin) * 2,
-    height: getSize(140),
-    marginTop: getSize(24),
-    borderColor: 'white',
-    borderWidth: getSize(1),
-    borderRadius: 20,
-  },
-  modalInputTitle: {
-    fontSize: getSize(14),
-    color: 'white',
-    fontFamily: 'Pretendard-SemiBold',
-    marginTop: getSize(18),
-  },
-  modalInput: {
-    marginRight: getSize(55),
-    marginBottom: getSize(25),
-    fontSize: getSize(14),
-    color: 'white',
-    fontFamily: 'Pretendard-Light',
-    marginTop: getSize(10),
-  },
 });
 
-export { DistanceUpdateModal, PaceUpdateModal, TimeUpdateModal };
+export {
+  DistanceUpdateModal,
+  PaceUpdateModal,
+  TimeUpdateModal,
+};
